+++ addons/bitrix_migration/tests/__init__.py
@@ -2,10 +2,10 @@ from . import test_role_mapping
 from . import test_markup_normalizer
 from . import test_attachments
 from . import test_departments
-from . import test_employee_avatars
 from . import test_meetings
 from . import test_meeting_comments
 from . import test_meeting_attachments
 from . import test_fired_employees
 from . import test_purge_noise
 from . import test_disk_filters
+from . import test_employee_hierarchy